requirements.txt
streamlit
pandas
numpy
requests
openai
python-dotenv
